requirements.txt
pygame==2.6.0
PyOpenGL==3.1.7
